Split client/server classes into separate header files
